==> 2020-432-book/15_ridgeandlasso.Rmd <==
# Other Variable Selection Strategies

## Why not use stepwise procedures?

1. The R^2^ for a model selected in a stepwise manner is biased, high.
2. The coefficient estimates and standard errors are biased.
3. The $p$ values for the individual-variable t tests are too small.
4. In stepwise analyses of prediction models, the final model represented noise 20-74\% of the time.
5. In stepwise analyses, the final model usually contained less than half of the actual number of real predictors.
6. It is not logical that a population regression coefficient would be exactly zero just because its estimate was not statistically significant.

This last comment applies to things like our "best subsets" approach as well as standard stepwise procedures.

Sander Greenland's comments on parsimony and stepwise approaches to model selection are worth addressing...

- Stepwise variable selection on confounders leaves important confounders uncontrolled.
- Shrinkage approaches (like ridge regression and the lasso) are far superior to variable selection.
- Variable selection does more damage to confidence interval widths than to point estimates.

If we are seriously concerned about **overfitting** - winding up with a model that doesn't perform well on new data - then stepwise approaches generally don't help.

@Vittinghoff2012 suggest four strategies for minimizing the chance of overfitting

1. Pre-specify well-motivated predictors and how to model them.
2. Eliminate predictors without using the outcome.
3. Use the outcome, but cross-validate the target measure of prediction error.
4. Use the outcome, and **shrink** the coefficient estimates.

The best subsets methods we have studied either include a variable or drop it from the model. 
Often, this choice is based on only a tiny difference in the quality of a fit to data. 

- @Harrell2001: not reasonable to assume that a population regression coefficient would be exactly zero just because it failed to meet a criterion for significance. 
- Brad Efron has suggested that a stepwise approach is "overly greedy, impulsively eliminating covariates which are correlated with other covariates."

So, what's the alternative?

## Ridge Regression

**Ridge regression** involves a more smooth transition between useful and not useful predictors which can be obtained by constraining the overall size of the regression coefficients.

Ridge regression assumes that the regression coefficients (after normalization) should not be very large. This is reasonable to assume when you have lots of predictors and you believe *many* of them have some effect on the outcome.

Pros:

1. Some nice statistical properties
2. Can be calculated using only standard least squares approaches, so it's been around for a while.
3. Available in the `MASS` package.


Ridge regression takes the sum of the squared estimated standardized regression coefficients and constrains that sum to only be as large as some value $k$.

\[
\sum \hat{\beta_j}^2 \leq k.
\]

The value $k$ is one of several available measures of the amount of shrinkage, but the main one used in the `MASS` package is a value $\lambda$. As $\lambda$ increases, the amount of shrinkage goes up, and $k$ goes down.

### Assessing a Ridge Regression Approach

We'll look at a plot produced by the `lm.ridge` function for a ridge regression for the prostate cancer study we worked on when studying Stepwise Regression and Best Subsets methods earlier.

- Several (here 101) different values for $\lambda$, our shrinkage parameter, will be tested.
- Results are plotted so that we see the coefficients across the various (standardized) predictors.
    + Each selection of a $\lambda$ value implies a different vector of covariate values across the predictors we are studying.
    + The idea is to pick a value of $\lambda$ for which the coefficients seem relatively stable.

```{r ridge_prost_code}
preds <- with(prost, cbind(lcavol, lweight, age, bph_f,
                           svi_f, lcp, gleason_f, pgg45))

x <- lm.ridge(prost$lpsa ~ preds, lambda = 0:100)
    
plot(x)
title("Ridge Regression for prost data")
abline(h = 0)
```

Usually, you need to use trial and error to decide the range of $\lambda$ to be tested. Here, `0:100` means going from 0 (no shrinkage) to 100 in steps of 1.

### The `lm.ridge` plot - where do coefficients stabilize?

Does $\lambda = 20$ seem like a stable spot here?

```{r ridge_prost_20}
x <- lm.ridge(prost$lpsa ~ preds, lambda = 0:100)
plot(x)
title("Ridge Regression for prost data")
abline(h = 0)
abline(v=20, lty=2, col="black")
```

The coefficients at $\lambda$ = 20 can be determined from the `lm.ridge` output. These are fully standardized coefficients. The original predictors are centered by their means and then scaled by their standard deviations and the outcome has also been centered, in these models.

```{r coeffs_at_lambda_20}
round(x$coef[,20],3)
```

Was an intercept used?

```{r intercept_at_lambda_20}
x$Inter
```

Yes, it was. There is an automated way to pick $\lambda$. Use the `select` function in the `MASS` package:

```{r select lambda}
MASS::select(x)
```

I'll use the GCV = generalized cross-validation to select $\lambda$ = 6 instead.

```{r c11_ridge_for_ptdsmale_with_40_line}
x <- lm.ridge(prost$lpsa ~ preds, lambda = 0:100)
plot(x)
title("Ridge Regression for prost data")
abline(h = 0)
abline(v=6, lty=2, col="black")

x$coef[,6]
```

### Ridge Regression: The Bottom Line

The main problem with ridge regression is that all it does is shrink the coefficient estimates, but it's not so useful in practical settings because it still includes all variables.

1. It's been easy to do ridge regression for many years, so you see it occasionally in the literature.
2. It leads to the **lasso**, which incorporates the positive features of shrinking regression coefficients with the ability to wisely select some variables to be eliminated from the predictor pool.

## The Lasso

The lasso works by takes the sum of the absolute values of the estimated standardized regression coefficients and constrains it to only be as large as some value k.

\[
\sum \hat{|\beta_j|} \leq k.
\]

This looks like a minor change, but it's not.

### Consequences of the Lasso Approach

1. In ridge regression, while the individual coefficients shrink and sometimes approach zero, they seldom reach zero and are thus excluded from the model. With the lasso, some coefficients do reach zero and thus, those predictors do drop out of the model. 
    + So the lasso leads to more parsimonious models than does ridge regression. 
    + Ridge regression is a method of shrinkage but not model selection. The lasso accomplishes both tasks.
2. If k is chosen to be too small, then the model may not capture important characteristics of the data. If k is too large, the model may over-fit the data in the sample and thus not represent the population of interest accurately.
3. The lasso is far more difficult computationally than ridge regression (the problem requires an algorithm called least angle regression published in 2004), although R has a library (`lars`) which can do the calculations pretty efficiently.

The lasso is not an acronym, but rather refers to cowboys using a rope to pull cattle from the herd, much as we will pull predictors from a model.

### How The Lasso Works

The `lars` package lets us compute the lasso coefficient estimates **and** do cross-validation to determine the appropriate amount of shrinkage. The main tool is a pair of graphs.

1. The first plot shows what coefficients get selected as we move from constraining all of the coefficients to zero (complete shrinkage) towards fewer constraints all the way up to ordinary least squares, showing which variables are included in the model at each point.
2. The second plot suggests where on the first plot we should look for a good model choice, according to a cross-validation approach.

```{r lasso_graph1_forprost}
## requires lars package
lasso1 <- lars(preds, prost$lpsa, type="lasso")
plot(lasso1)
```

- The y axis shows standardized regression coefficients. 
    + The `lars` package standardizes all variables so the shrinkage doesn't penalize some coefficients because of their scale. 
- The x-axis is labeled `|beta|/max|beta|`. 
    + This ranges from 0 to 1.
    + 0 means that the sum of the $|\hat{\beta_j}|$ is zero (completely shrunk)
    + 1 means the ordinary least squares unbiased estimates.
    
The lasso graph starts at constraining all of the coefficients to zero, and then moves toward ordinary least squares. 

Identifiers for the predictors (numbers) are shown to the right of the graph.

The vertical lines in the lasso plot show when a variable has been eliminated from the model, and in fact these are the only points that are actually shown in the default lasso graph. The labels on the top of the graph tell you how many predictors are in the model at that stage.

```{r summ_lasso_1}
summary(lasso1)
```

Based on the $C_p$ statistics, it looks like the improvements continue throughout, and don't really finish happening until we get pretty close to the full model with 9 df.

### Cross-Validation with the Lasso

Normally, cross-validation methods are used to determine how much shrinkage should be used. We'll use the `cv.lars` function.

- 10-fold (K = 10) cross-validation
    + the data are randomly divided into 10 groups. 
    + Nine groups are used to predict the remaining group for each group in turn.
    + Overall prediction performance is computed, and the machine calculates a cross-validation criterion (mean squared error) and standard error for that criterion.
    
The cross-validation plot is the second lasso plot.

```{r lasso_graph2}
set.seed(432)
lassocv <- cv.lars(preds, prost$lpsa, K=10)
## default cv.lars K is 10
```

We're looking to minimize cross-validated mean squared error in this plot, which doesn't seem to happen until the fraction gets very close to 1.

### What value of the key fraction minimizes cross-validated MSE?

```{r}
frac <- lassocv$index[which.min(lassocv$cv)]
frac
```

The cross-validation plot suggests we use a fraction of nearly 1.0, suggesting that all of the predictors will be kept in, based on the top LASSO plot.

```{r lasso_bothplots, fig.height=7}
par(mfrow=c(2,1))
lasso1 <- lars(preds, prost$lpsa, type="lasso")
plot(lasso1)
set.seed(432)
lassocv <- cv.lars(preds, prost$lpsa, K=10)
par(mfrow=c(1,1))
```

### Coefficients for the Model Identified by the Cross-Validation

```{r}
coef.cv <- coef(lasso1, s=frac, mode="fraction")
round(coef.cv,4)
```

So the model suggested by the lasso still includes all eight of these predictors.

### Obtaining Fitted Values from Lasso

```{r}
fits.cv <- predict.lars(lasso1, preds, s=frac, 
                        type="fit", mode="fraction")
fits.cv
```

### Complete Set of Fitted Values from the Lasso

```{r lasso_fits_rounded}
round(fits.cv$fit,3)
```

To assess the quality of these predictions, we might plot them against the observed values of our outcome (`lpsa`), or we might look at residuals vs. these fitted values.

```{r}
prost_lasso_res <- data_frame(fitted = fits.cv$fit, 
                             actual = prost$lpsa, 
                             resid = actual - fitted)

ggplot(prost_lasso_res, aes(x = actual, y = fitted)) + 
    geom_point() + 
    geom_abline(slope = 1, intercept = 0) +
    labs(y = "Fitted log(PSA) from Cross-Validated LASSO",
         x = "Observed values of log(PSA)",
         title = "Fitted vs. Actual Values of log(PSA)")
```

```{r}
ggplot(prost_lasso_res, aes(x = fitted, y = resid)) + 
    geom_point() + 
    geom_hline(yintercept = 0, col = "red") +
    geom_smooth(method = "loess", col = "blue", se = F) +
    labs(x = "LASSO-fitted log(PSA)",
         y = "Residuals from Cross-Validated LASSO model",
         title = "Residuals vs. Fitted Values of log(PSA) from LASSO",
         subtitle = "with loess smooth")
```

### When is the Lasso Most Useful?

As @Faraway2015 suggests, the lasso is particularly useful when we believe the effects are sparse, in the sense that we believe that few of the many predictors we are evaluating have a meaningful effect. 

Consider, for instance, the analysis of gene expression data, where we have good reason to believe that only a small number of genes have an influence on our response of interest.

Or, in medical claims data, where we can have thousands of available codes to search through that may apply to some of the people included in a large analysis relating health care costs to outcomes.

## Applying the Lasso to the `pollution` data

Let's consider the lasso approach in application to the `pollution` data we've seen previously. Recall that we have 60 observations on an outcome, `y`, and 15 predictors, labeled x1 through x15.

```{r lasso_graph1_forpollution}
preds <- with(pollution, cbind(x1, x2, x3, x4, x5, x6, x7,
                               x8, x9, x10, x11, x12, x13,
                               x14, x15))

lasso_p1 <- lars(preds, pollution$y, type="lasso")
plot(lasso_p1)
```

```{r summ_lasso_p1}
summary(lasso_p1)
```

Based on the $C_p$ statistics, it looks like the big improvements occur somewhere around the move from 6 to 7 df. Let's look at the cross-validation.

```{r}
set.seed(432012)
pollution_lassocv <- cv.lars(preds, pollution$y, K=10)
```

Here it looks like cross-validated MSE happens somewhere between a fraction of 0.2 and 0.4.

```{r}
frac <- pollution_lassocv$index[which.min(pollution_lassocv$cv)]
frac
```

```{r, fig.height=7}
par(mfrow=c(2,1))
lasso_p1 <- lars(preds, pollution$y, type="lasso")
plot(lasso_p1)
set.seed(432012)
pollution_lassocv <- cv.lars(preds, pollution$y, K=10)
par(mfrow=c(1,1))
```

It looks like a model with 6-8 predictors will be the most useful. The cross-validated coefficients are as follows:

```{r}
poll.cv <- coef(lasso_p1, s=frac, mode="fraction")
round(poll.cv,3)
```

Note that by this cross-validated lasso selection, not only are the coefficients for the 8 variables remaining in the model shrunken, but variables `x4`, `x5`, `x10`, `x11`, `x12`, `x13` and `x15` are all dropped from the model, and model `x8` almost is, as well.

```{r lasso_fit1}
poll_fits <- predict.lars(lasso_p1, preds, s=frac, 
                        type="fit", mode="fraction")
round(poll_fits$fit,3)
```

Here's a plot of the actual `pollution` y values, against these fitted values.

```{r}
poll_lasso_res <- data_frame(fitted = poll_fits$fit, 
                             actual = pollution$y, 
                             resid = actual - fitted)

ggplot(poll_lasso_res, aes(x = actual, y = fitted)) + 
    geom_point() + 
    geom_abline(slope = 1, intercept = 0) +
    labs(y = "Fitted y values from Cross-Validated LASSO",
         x = "Observed values of y = Age-Adjusted Mortality Rate",
         title = "Fitted vs. Actual Values of Age-Adjusted Mortality")
```

And now, here's a plot or residuals vs. fitted values.

```{r}
ggplot(poll_lasso_res, aes(x = fitted, y = resid)) + 
    geom_point() + 
    geom_hline(yintercept = 0, col = "red") +
    geom_smooth(method = "loess", col = "blue", se = F) +
    labs(x = "LASSO-fitted Age-Adjusted Mortality",
         y = "Residuals from Cross-Validated LASSO model",
         title = "Residuals vs. Fitted Values of Age-Adjusted Mortality from LASSO",
         subtitle = "with loess smooth")
```
